pyproject.toml
[tool.poetry]
name = "atlecta-api"
version = "0.1.0"
description = ""
authors = ["David Lashevich <david.lashevich@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.110.1"
alembic = "^1.13.1"
sqlalchemy = "^2.0.29"
uvicorn = "^0.29.0"
asyncpg = "^0.29.0"
python-dotenv = "^1.0.1"
fastapi-users-db-sqlalchemy = "^6.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
